refactor: Accommodation 관련 도메인 클래스들 domain 패키지로 이동
