+++ backend/src/main/java/codesquad/airbnb/accomodation/domain/Address.java
@@ -1,4 +1,4 @@
-package codesquad.airbnb.accomodation;
+package codesquad.airbnb.accomodation.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
